+++ scripts/test/PostMaster/ExternalTicketNumberRecognition.t
@@ -440,6 +440,35 @@ This is an update to incident by mail',
     },
     {
         Name =>
+            '#12 - Added option CcAddressRegExp - matching Cc',
+        Email => 'From: Sender <sender@example.com>
+To: Some Name <recipient@example2.com>
+Cc: Some Other Name <recipient@example3.com>
+Subject: An incident subject ' . $ExternalTicketID . '
+
+This is an update to incident by mail',
+        Check => {
+            "DynamicField_$FieldName" => $ExternalTicketID,
+        },
+        JobConfig => {
+            DynamicFieldName  => $FieldName,
+            FromAddressRegExp => '\\s*@example.com',
+            ToAddressRegExp   => '\\s*@example2.com',
+            CcAddressRegExp   => '\\s*@example3.com',
+            Module            => 'Kernel::System::PostMaster::Filter::ExternalTicketNumberRecognition',
+            Name              => 'Some Description',
+
+            NumberRegExp         => '(?:Some test\\s|incident subject\\s)(\\d.*)',
+            SearchInBody         => '0',
+            SearchInSubject      => '1',
+            SenderType           => 'system',
+            IsVisibleForCustomer => 1,
+            TicketStateTypes     => 'new;open',
+        },
+        NewTicket => 2,
+    },
+    {
+        Name =>
             '#13 - Added option ToAddressRegExp - non-matching To',
         Email => 'From: Sender <sender@example.com>
 To: Some Name <recipient@example999.com>
@@ -453,6 +482,35 @@ This is an update to incident by mail',
             DynamicFieldName  => $FieldName,
             FromAddressRegExp => '\\s*@example.com',
             ToAddressRegExp   => '\\s*@example2.com',
+            Module            => 'Kernel::System::PostMaster::Filter::ExternalTicketNumberRecognition',
+            Name              => 'Some Description',
+
+            NumberRegExp         => '(?:Some test\\s|incident subject\\s)(\\d.*)',
+            SearchInBody         => '0',
+            SearchInSubject      => '1',
+            SenderType           => 'system',
+            IsVisibleForCustomer => 1,
+            TicketStateTypes     => 'new;open',
+        },
+        NewTicket => 1,
+    },
+    {
+        Name =>
+            '#13 - Added option CcAddressRegExp - non-matching Cc',
+        Email => 'From: Sender <sender@example.com>
+To: Some Name <recipient@example2.com>
+Cc: Some Name <recipient@example999.com>
+Subject: An incident subject ' . $ExternalTicketID . '
+
+This is an update to incident by mail',
+        Check => {
+            "DynamicField_$FieldName" => undef,
+        },
+        JobConfig => {
+            DynamicFieldName  => $FieldName,
+            FromAddressRegExp => '\\s*@example.com',
+            ToAddressRegExp   => '\\s*@example2.com',
+            CcAddressRegExp   => '\\s*@example3.com',
             Module            => 'Kernel::System::PostMaster::Filter::ExternalTicketNumberRecognition',
             Name              => 'Some Description',